[All] Move RegisterRegistryPasses to PimcoreBundle in order to fix dependency issue
