+++ src/CoreShop/Bundle/AddressBundle/DependencyInjection/Compiler/CompositeCountryContextPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\AddressBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
 
 final class CompositeCountryContextPass extends PrioritizedCompositeServicePass
 {
