+++ src/CoreShop/Bundle/IndexBundle/DependencyInjection/Compiler/RegisterGetterPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\IndexBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
 
 class RegisterGetterPass extends RegisterRegistryTypePass
 {
