+++ src/CoreShop/Bundle/LocaleBundle/DependencyInjection/Compiler/CompositeLocaleContextPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\LocaleBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
 
 final class CompositeLocaleContextPass extends PrioritizedCompositeServicePass
 {
