+++ src/CoreShop/Bundle/OrderBundle/DependencyInjection/Compiler/CartPriceRuleConditionPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterRegistryTypePass;
 
 final class CartPriceRuleConditionPass extends RegisterRegistryTypePass
 {
