+++ src/CoreShop/Bundle/OrderBundle/DependencyInjection/Compiler/RegisterCartContextsPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass;
 
 final class RegisterCartContextsPass extends PrioritizedCompositeServicePass
 {
