+++ src/CoreShop/Bundle/PimcoreBundle/DependencyInjection/Compiler/RegisterGridFilterPass.php
@@ -12,8 +12,6 @@
 
 namespace CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
-
 class RegisterGridFilterPass extends RegisterSimpleRegistryTypePass
 {
     public function __construct()
