+++ src/CoreShop/Bundle/ResourceBundle/DependencyInjection/Compiler/RegisterSimpleRegistryTypePass.php
@@ -12,61 +12,12 @@
 
 namespace CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass as BaseRegisterSimpleRegistryTypePass;
 
-abstract class RegisterSimpleRegistryTypePass implements CompilerPassInterface
+/**
+ * @deprecated Deprecated since 2.0.0-beta.2 and will be removed in 2.0.0-beta.3, use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\PrioritizedCompositeServicePass instead
+ */
+abstract class RegisterSimpleRegistryTypePass extends BaseRegisterSimpleRegistryTypePass
 {
-    /**
-     * @var string
-     */
-    protected $registry;
 
-    /**
-     * @var string
-     */
-    protected $parameter;
-
-    /**
-     * @var string
-     */
-    protected $tag;
-
-    /**
-     * @param string $registry
-     * @param string $parameter
-     * @param string $tag
-     */
-    public function __construct($registry, $parameter, $tag)
-    {
-        $this->registry = $registry;
-        $this->parameter = $parameter;
-        $this->tag = $tag;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function process(ContainerBuilder $container)
-    {
-        if (!$container->has($this->registry)) {
-            return;
-        }
-
-        $registry = $container->getDefinition($this->registry);
-
-        $map = [];
-        foreach ($container->findTaggedServiceIds($this->tag) as $id => $attributes) {
-            if (!isset($attributes[0]['type'])) {
-                throw new \InvalidArgumentException('Tagged Service `'.$id.'` needs to have `type` attribute.');
-            }
-
-            $map[$attributes[0]['type']] = $attributes[0]['type'];
-
-            $registry->addMethodCall('register', [$attributes[0]['type'], new Reference($id)]);
-        }
-
-        $container->setParameter($this->parameter, $map);
-    }
 }
