+++ src/CoreShop/Bundle/TrackingBundle/DependencyInjection/Compiler/TrackerPass.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\TrackingBundle\DependencyInjection\Compiler;
 
-use CoreShop\Bundle\ResourceBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
+use CoreShop\Bundle\PimcoreBundle\DependencyInjection\Compiler\RegisterSimpleRegistryTypePass;
 
 final class TrackerPass extends RegisterSimpleRegistryTypePass
 {
